app/requirements.txt
fastapi
uvicorn[standard]
pydantic
openai
sentence-transformers
faiss-cpu
numpy